lory/pyproject.toml
[project]
name = "lory"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.7",
    "djangorestframework>=3.16.1",
]

[dependency-groups]
dev = [
    "django-debug-toolbar>=6.0.0",
]
prod = [
    "gunicorn>=23.0.0",
]
